Adds a few extra engineering crates (#2786)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds the following crates:
- Turbine Crate: Contains a power turbine, power compressor, and turbine
computer board (4000 credits)
- Holofield Generator Crate: Contains two holofield generator boards
(1000 credits)
- Ion Thruster Crate: Contains an ion thruster and an engine precharger
board (1500 credits)
- Plasma Thruster Crate: Contains a plasma thruster and an engine heater
board (1500 credits)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This permits ships without RnD to modify or repair themselves in the
wake of RnD being cut from most places - holofield generators and
thrusters are on basically every vessel, so it's not as though
cutting-edge research is needed to make them. As for the turbine, making
it available as a slightly-cheaper alternative to the TEG (especially
with its rework which lets it produce thrust as well) for ships which
can manage to source the plasma gas to fuel it.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added turbine, holofield generator, ion thruster, and plasma
thruster crates
/🆑

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->