Removes the singularity, tesla and particle accelerator packs from cargo (#2787)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

The singularity, tesla, and particle accelerator can no longer be bought
since the engines are either unbalanced or impossible to effectively use
on most ships

## Why It's Good For The Game

The singularity can't be effectively fitted to any ship without building
in hyperspace. Even if it is, it's going to irradiate/EMP everything in
a large area, which likely means somewhere important

The tesla is just free infinite power if set up, which is boring. It
doesn't have any upkeep and containment is extremely simple

The particle accelerator doesn't do anything except turn on and power
those two engines

## Changelog

:cl:
del: The tesla, singularity, and particle generator are no longer
purchasable from cargo
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Signed-off-by: Theos <theubernyan@gmail.com>