Emergency Response Team Housekeeping (#2744)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR introduces a system to limit special roles picked in ERTs, which
can be toggled in the settings screen, and sorts out the antagonist
datums into factional files.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/e3e18d92-6dfc-4be6-8d0f-7bf1bc366b3a)
This could happen to you.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
code: ERT antag datums are now sorted into factional files
admin: ERTs can have limited special roles
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: meem <75212565+meemofcourse@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>