Tweaks 38 damage and adds .38 Surplus (#2775)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Reduces .38 damage (Detective Special, Flaming Arrow) back down to 20
per shot at -20 AP, identical to 9mm, making it not just .45 Except You
Can Print The Normal Version On A Lathe, and adds a surplus variant that
is printed on the lathe.

Yes, I'm well aware this is a pretty substantial nerf; if you're using
.38 Surplus, as many ships will end up doing, you're doing 10 less
damage per shot than before this PR, at a slower rate of fire and with
smaller magazine sizes than 9mm guns. Outpost resupply for .38 should be
quite cheap to compensate. Would be nice to make the .38 ammo variants a
little more obtainable, too. Future PR material.

## Why It's Good For The Game

I'm not sure why this was buffed since it just counterintuitively made
.38 a clone of .45 (at least making it a clone of 9mm makes more sense
with the cartridge's irl properties), and surplus .38 SPC ammo has been
needed for a long time.

## Changelog

:cl:
add: .38 Special Surplus ammo is now in the autolathe.
balance: Nerfed .38's damage back down to 20
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->