Better Food Crates (#2801)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds more variety to the food crates, allowing enterprising chefs to
order ingredients that can actually make a modest amount of food instead
of needing to make a full botany setup. It's still a bit sparse, and one
day I want to be able to order specific ingredients for a higher price,
but for now, this is an improvement I think.

I'm willing to discuss the costs of the crates. Money is still so
abstract that I really want input on what this amount of stuff should
cost.

## Why It's Good For The Game
Actually allows chefs to cook a reasonable amount of food with
ingredients purchased from the outpost, instead of just having to make a
full botany setup.

## Changelog

:cl:
tweak: Makes the basics ingredients crate less barren and useless.
add: Advanced cooking crate, for buying the condiments you only need a
little bit of.
add: Standard meat crate, when you don't want to buy the weird exotic
meat crate.
add: Grains crate, for when you want to make a lot of flour for cooking
add: Bread crate, for when you're lazy and want bread for recipes. Or to
eat. I don't care.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->