fixes 8x50 boxes and also adds them and some other ammo to the elder (#2818)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
sorry for not atomizing more i have a really bad tummyache today
## Why It's Good For The Game
boxes being invisible and having 50 rounds is bad and also its good to
like have ammo in the hunting. ship

## Changelog

:cl:
add: More and appropriate ammo to the Elder-class.
fix: Fixed a few oversights with 8x50 ammo boxes.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->