Adds more ways to ignite e-cigarettes. (#2796)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Do you want your character to vape, but did they not spawn with a PDA?
Well, worry not, this PR adds new ways to ignite your
e-cigarettes/e-cigars!

First one, is with a tablet. Works essentially the same as with a PDA.
There is way more jobs that get a free tablet than those that get a PDA.
Plus, you can get those from loadout.

Second one, is with an APC. Smash your vape on the APC and it will get
ignited! How does it work? Clearly designers were just very thoughtful
and included a vape igniter or whatever you use!

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Items being actually usable is nice.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: E-cigarettes and e-cigars can now be turned on with tablets and
APCs.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->