missing icon unit test  (#2530)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

ports tests related to icons from tg
a lot of missing sprites
while most of the "missing" sprites don't really matter for multiple
reasons:

- Some sprites simply overwrite there icon_state on inits and alot
completely ignore the icon like with limbs
- Others are completely unused/unobtainble content like old Chaplin
- A shit ton of items declare items states but dont have a sprite to
match. For nearly all of these you can just remove the line as not every
item needs a inhand state
- Items that equip to the belt slot but dont have a icon in the belt dmi
to match there icon_state fail. We have a seperate var specificly for
declaring the icon for wearing it we can fix most of these with.

the tests should help for debugging future prs once we get all of the
current sprites fixed
If anyone wants to help go through and fix em it would really help
because there is over 1000 fails at this time of writing

For fixing sprites that aren't important im going to skim through our
current sprites and tgs but I dont have the time to make new ones.

Should fix #2619 and #1275

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

helps find missing icons?
fixes a shit ton of sprites

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
code: various missing icon unit tests
fix: to many sprites to name
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>