Economy logging + cruft trimming (#2684)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds a bit of logging to see how money is flowing through the in-game
economy to base further changes on, as well as removed a lot of old,
unmaintained, and/or unneeded features.

## Why It's Good For The Game

We want to make the economy more engaging, so we need to find out where
the worst offenders of getting loads of money is from.

## Changelog

:cl:
tweak: Medical kiosks no longer charge for their service
admin: The economy subsystem now tracks how much money is currently
in-round
admin: Adds feedback for the movement of money
del: FULLY removes departmental budgets
del: Removes security record citations
del: "ancap" mode
del: "egalitarian" mode
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Mark Suckerberg <mark@suckerberg.gay>