Gives the Riggs a touch-up (#2721)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![rigger_remap](https://github.com/shiptest-ss13/Shiptest/assets/60533805/53fce13e-7ed8-46de-a878-077bf0614077)
The Riggs has aged incredibly well for being one of the oldest maps
still in rotation, but it's definitely fallen out of line with newer
ships as standards have been updated. This PR brings it in line with
those standards and freshens up the layout and decoration a bit.

- Removed RnD and ORM
- Added separate cryo room
- Added crew lounge area and miner ready room
- Moved atmos slightly to make accessing the empty rooms easier and
replaced the standing air tanks with a 1-tile open air tank (temporary,
really, until we get better tanks coded)
- Added access restrictions to atmos and engine room (but not
engineering bay)
- Reduced miner slots to 1 and assistant slots to 2
- Removed Mech Technician

## Why It's Good For The Game

Riggs was getting crusty and needed some love.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Made the Riggs-class sloop nicer to live in
balance: Removed RnD and ORM from the Riggs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->