Fixes 38 Surplus being unprintable in the autolathe and 38 speedloader description (#2840)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Fixes #2814 and makes 38 surplus ammo able to be printed from the
autolathe. Changes the damage blurb on the 38 speedloader to reflect
standard 38 ammo's new stats.
## Why It's Good For The Game
Bug fixes are good.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: 38 Surplus ammo should be printable in the autolathe.
fix: 38 speedloader damage blurb is accurate to it's new stats.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->