The ShipCut (#2870)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Removes the BYO, the Halftrack, the Masinyane, the PillBottle, the Rube
Goldberg, the Shepherd, the Nemo, the Radio_Funny, and the Lugol

Aspawns the Lagoon and Junker

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
As Maptainers we came together and said "Yeah". These ships are not up
to our standards and will bring up gameplay standards by the lack of
their presence.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:phoaly
del: Removed some old ships
tweak:  Aspawned the Lagoon and Junker
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->