Adds Fullscreen (#2878)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Added fullscreen as hotkey F11 and on OOC -> Toggle Fullscreen

![image](https://github.com/shiptest-ss13/Shiptest/assets/58084138/3db51657-c491-423a-a4e2-1e70647897b3)

![image](https://github.com/shiptest-ss13/Shiptest/assets/58084138/8136e422-922a-4937-8c59-50993d1a8bc1)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
You can play in fullscreen without annoying borders.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Fullscreen on F11
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->