Constructable exercise equipment (#2839)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Punching bags can now be constructed from 5 cloth and deconstructed
with a wirecutter.
- Bench presses and chest presses can be constructed from 5 metal and
deconstructed with a wrench.
## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
You no longer need to smash apart the exercise equipment like a caveman,
and offers players some more freedom in how they want to decorate or
arrange their ships.


## Changelog

:cl:
add: Punching bags can be built with 5 cloth.
add: Bench press and chest press can be built with 5 metal.
add: Exercise equipment can be deconstructed.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->