Clears up the graffiti list, adds hobocode graffiti. (#2887)
## About The Pull Request
This PR:
Removes these graffiti sprites.

![73a8f5806f023f6444d442b37162fced](https://github.com/shiptest-ss13/Shiptest/assets/118859017/866e500a-ad54-4209-b33f-547444b1e0e0)

Adds a new set of codersprite graffiti, which is based on hobo symbols. 

![74bd1076299c63cafd7822ced40010e5](https://github.com/shiptest-ss13/Shiptest/assets/118859017/ce57deb2-b42a-4e0e-8cd3-7b37b924267b)

![image](https://github.com/shiptest-ss13/Shiptest/assets/118859017/6b8ea581-b26e-4cc4-9e40-93c6207973de)

## Why It's Good For The Game
We've got a bunch of cruft in spraycan graffiti, including a ton of
weird anti-furry stuff, and references that don't make sense in our
setting. This PR removes the bad stuff, and adds some new graffiti to
fill the space.

## Changelog

🆑
add: A new graffiti section for code symbols has been added.
del: Some graffiti has been removed from the spraycan list.
/🆑

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->