Adjusted Pure Soulus Dust (#2935)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adjusts pure soulus dust to not require system cleaner, and output is
20u. It already requires 2 very hard to get reagents on shiptest
anyways.

## Why It's Good For The Game

It's a nice chem, but holy/hollow water are super duper rare, on top of
needing to collect legion cores. It doesn't need to be this hard. In a
hilarious sense...... straight applying legion cores is better pre-PR

## Changelog

:cl:
balance: pure soulus dust now makes 20u per reaction, and system cleaner
is not required.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->