Adds storage to helmets, which can overlay certain items in them (#2845)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/cc1a9bfc-5f6a-4091-8ea5-695c5b1e1ca7)
Adds a whitelisted storage component to helmets, and a variable that
determines if they should process overlays. Features codersprites.

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/8e874d88-885b-4909-9f5d-81be1ae4f5ae)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
  * It looks nice.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Helmets now have a small, two-item storage. You can put goggles in
there, and it might overlay them.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->