Fixes stray pixels on north-facing siding ends. (#2985)
## About The Pull Request

![image](https://github.com/shiptest-ss13/Shiptest/assets/118859017/738ecb6b-45b2-4114-bfd9-b4089664013b)
This is killing me.

## Why It's Good For The Game

I will continue to wither until this is merged.

## Changelog

🆑
fix: Fixed stray pixels on north-facing sidings.
/🆑

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->