Cuts the clown, mime, deity and religion names and loadout bibles into little itty pieces. (#2966)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes cruft!!!! finally!! 

## Why It's Good For The Game

im like doomguy except instead of killing demons im killing the ancient
ghosts of tgstation humor
## Changelog

:cl:

del: Removed the bible from loadout, as well as the religion and deity
names because god is dead. Clown and mime names too.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->