Black Market Revamp (#2929)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

### Mechanical stuff

- Removes Teleportation as a shipping option from the Black Market. The
code behind it was removed so it only ate your money.
- Changes how the LTSRBT handles orders. The uplink is now linked to the
LTSRBT as it's destination instead of checking for a free telepad across
the entire world.
- Adds examines for the Black Market uplink, changes shipping
descriptions to be more clear
- Adds the pair_item var for black market items. If the main item
spawns, it's paired item is guaranteed to spawn alongside it.
 - The LTSRBT gets orders 10x faster.
- The Black Market Uplink now uses a microlaser instead of an ansible in
it's crafting recipe.
-  A sawn off Illestren typepath
### Catalogue

- Reduces the cost of the LTSRBT at the outpost from 5000 to 1000.
- Adds a whole bunch of new items to the Black Market catalogues, and
changes the prices of existing items to be in line with the current
economy.
- Removes items and adjusted some descriptions from the Black Market
that didn't fit Shiptest's setting.
- Adds two new tabs to the Black Market, Explosives and Emergency.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Honestly, this started as a redo of #2927 to add Eoehoma weapons to the
Black Market instead, but the Black Market was kinda shit. The teleport
option was broken, the black market uplink could send to any LTSRBT that
existed, so someone half the sector away could end up with your orders
(if anyone actually bought them that is) and the item selection was
pretty poor and still based around a station economy price wise.

Clears up the cruft, and adds a bunch of new items to the Black Market
catalogue to hopefully make it an appealing alternative option to buying
from the outpost. It's a thematic way to sell items normally wouldn't be
seen. Most of the stock is limited as well, which adds a sense of
urgency and competition with other ships if there's an item that catches
your eye.

Dropped the price of the LTSRBT and changed the uplink recipe to make
the Black Market more accessible. You weren't finding an ansible unless
you have RND or got lucky. The LTSRBT was also comedically slow to the
point where you were left wondering if it worked.

The sawn off Illestren was so I could directly spawn it for the Black
Market. I'd thought it'd be funny.
 
 I'm happy to change any of the items, prices or stocks if needed. 

## Changelog

:cl:
add: A bunch of new items for the Black Market catalogue
add: Two new Black Market Tabs, Explosives and Emergency
add: Sawn off Illestren typepath
del: Black Market Teleportation
del: Space Ninja mask and Clown Tears from Black Market
tweak: Black Market Uplink uses a micro laser instead of an ansible for
crafting
tweak: Black Market descriptions and examines
tweak: Black Market Uplinks link directly to a specific LTSRBT
balance: Black Market item price and stock adjusted
balance: LTSRBT is faster
code: pair_item var for Black Market items
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Gristlebee <56049844+Gristlebee@users.noreply.github.com>
Co-authored-by: meem <75212565+meemofcourse@users.noreply.github.com>