Tortuga-class Converted Cargo Freighter (#2765)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request


![tortuga](https://github.com/shiptest-ss13/Shiptest/assets/75212565/981781e4-d57c-4f21-93ad-496e3ed5f883)

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/252de101-ee3d-4e3b-a0e4-6475f8fadec7)

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/151c1302-9e32-42db-83af-97db27d8dc87)

![imagen](https://github.com/shiptest-ss13/Shiptest/assets/75212565/d4b82622-406c-45a0-9c1e-1781b2802cf4)



Originally a cargo transport vessel, the Tortuga class is an improvised
combat vessel under the ranks of the Frontiersmen fleet. Clever field
modifications include the conversion of the cargo bay to a hangar, and
additional padding around the ship in the form of asteroid rocks. It
serves a dual purpose, both as a capable ship for establishing a
military enclave in a sector, and as a sturdy vessel for reconnaissance.
The nimble Brawler-class subshuttle of the Tortuga is offset by its very
low overall mobility and loss of operation if the subshuttle were to be
lost.

It is staffed by:
* 1 Captain
* 1 Shipswain (Head of Security)
* 1 Aidman
* 1 Sapper
* 2 Boarders
* 2 Rookies

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The Ember is a pseudo-dreadnaught, and unfortunately, it kind of sucks.
Pirates need a subshuttle to effectively raid.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Tortuga-class.
add: Brawler-class.
del: Ember-class.
del: Scar suit.
tweak: Frontiersmen outfits start with less stuff.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: meem <75212565+meemofcourse@users.noreply.github.com>