Remaps the Lavaland Winter Biodome (#2856)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
### Extrasolar Biodome Project

The Most Serene Solar and Intersolar Confederation welcomes you to one
of our many experimental biodomes out in the frontier.
These facilities are developed and built as a collaboration with the
Pionierskompanien and are to serve as extrasolar research bases for our
most accomplished scientists and scribes. These facilities utilize
breakthroughs in terraforming technology to create a comfortable,
habitable environment in even the harshest of locales. Your purpose here
is to research and report on the effectiveness of the technology
utilized within the central dome, as well as maintain and protect the
facility.

This PR remaps the lavaland winter biodome into a facility used by the
Pionierskompanien to field-test terraforming and environmental
regulation technologies. Unfortunately for the occupants, it seems the
local fauna had other ideas.

The main loot of this ruin is: 5 Plasma sheets, a small sum of credits,
a Pistole C and a Claris in the display case.

### Pictures

![StrongDMM-2024-03-23 21 57
03](https://github.com/shiptest-ss13/Shiptest/assets/95449138/44160bb2-900c-4a8c-b002-0b1854c5c2b6)
![StrongDMM-2024-03-23 21 57
04](https://github.com/shiptest-ss13/Shiptest/assets/95449138/d63272ba-7b12-43f3-90a0-ffeca8bf4130)

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/cd8de5cc-68bf-49b3-9179-864efa8c4753)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Fixes yet another crufty TG ruin that had some
ever-so-racially-insensitive mobs within it. Those have been removed, by
the way.
This also places SolCon into the ruin pool, hopefully allowing for me to
make a series out of similar projects like this on other planets.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Remaps lavaland_surface_biodome_winter.dmm
del: Removes the "Undead Eskimo" mob
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: zevo <95449138+Zevotech@users.noreply.github.com>