Makes Plastic Flaps no longer airtight (#2969)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
# You cannot tell me these fucking things are blocking the vacuum of
space.

![Baggage-Claim-Strip-Curtain-Kit-Shop-All-StripCurtaincom_500__00450](https://github.com/shiptest-ss13/Shiptest/assets/95449138/e9397556-db2f-4fac-aa05-5c4fa490ec93)
(Makes plastic flaps no longer block atmos and changes the name and desc
to match.)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Players should not be able to construct a budget-tinyfan for 5 plastic
with only the caveat of having to crawl under it. Using these to cover
bay doors has been frowned upon for a while and it should have been
fixed when tinyfans were phased out.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Plastic flaps no longer block atmos
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->