Blastdoor and Atmos Holofield ID multitool syncing (#3004)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Multitools can be used on a button with it's panel open to add a blast
door or holofield controller to it's buffer.

The ID can be copied onto a blast door the same way as changing it
normally with a multitool. For holofields, you need to have the access
mode on unlocked and it's panel closed.

Adds holofield controllers to the autolathe.

Changes all the span classes to span macros

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

When you have a roundstart blast door break or holofield break for
whatever reason, it's almost impossible to replace and properly resync
without ripping out the rest of the blastdoors or the other holofield.
This should make reconstruction more convienent.

Also being able to make and use your own holofield controller adds a
more intuitive way to rig up player made holofields, as currently you
have to use signallers on the activation wire.

## Changelog

:cl:
add: Autolathe can print holofield controllers.
add: You can sync blastdoors and holofields to a controller in a button
with a multitool
code: span macros for blast doors, buttons and holofields
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Gristlebee <56049844+Gristlebee@users.noreply.github.com>
Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>