The Mechening: Shiptest Edition (#3035)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Hello there! Hope you didn't miss me this soon ;)

This PR ports a very recent PR on /tg/, done by the extremely talented
INFRARED_BARON and PR'd by @githubuser4141 (known as wastelandmain on
discord)
- https://github.com/tgstation/tgstation/pull/83024

Since we do have a few original mechs on Shiptest as well, the CLIP
mechs have also been resprited entirely to match the new suits we will
now see in game!


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/3b03bbc3-a1ba-4403-9e3b-aa9c84c21b35)

![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/1df55795-5f6b-4dfe-a5b7-170eea7b24d4)

### This PR has been ported with the consent of the original PR author,
as well as the original spriter.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
They're a lot prettier than what we currently have, and while we wait
for the fabled mech rework, this is a good enough visual stopgap to not
make me want to tear my eyeballs out.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: INFRARED_BARON, PositiveEntropy
imageadd: Resprites nearly every mech!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->