Bandoliers now hold 40 items as intended (#3014)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Bandolier max items was set to 40 but max weight was not, meaning it
only held up to 14 total weight which is significantly less than 40

## Why It's Good For The Game

Fix

## Changelog

:cl:
fix: bandoliers now hold the correct amount of bullets (several more)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Signed-off-by: Theos <theubernyan@gmail.com>