Fixes create and destory failure due to missing physiology (#2999)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Sometimes during unit tests (prob shouldn't happen in the first place) a
human will get created and have electrocute act called by a tesla
anomaly which causes a fail because it looks for a shock resist in
physiology which doesn't yet exist.
If humans are shocked during init and it does not have its physiology
yet it now returns
It may be better to just have an if physiology for fetching the
physiology siemens but if its missing its physiology its prob missing
other thing needed like generating organs or dna
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
im tired of half my prs failing from this

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: fixed create and destroy test fail
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->