Adds a salve that allows elzuose to change color (#2995)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
To be merged after #2993.
Adds a Kalixcian beauty product that allows Elzuose to (temporarily (for
the duration of the round or until used again)) change their glow color.
Available in the loadout next to hair dye.


https://github.com/shiptest-ss13/Shiptest/assets/95449138/01060ea1-4b6b-490f-9bf2-83b69cf7efd7


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People wanted to keep the ability to change the glow color of an elzuose
using an item. Fits better than just using a multitool without the
potential for grief/shenanigans. Lore approved too!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Adds the Elzuose color salve, a product that allows Elzuose to
change their glow color upon application.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->