makes plant analyzer usable (#3054)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

### Tray
#### New

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/d8000903-a18b-460c-a9b5-716b3416a98f)
#### Old

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/5b16e40a-e12b-42e2-a1e1-324b702ea572)


### Plant
#### New

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/22bb48bd-9485-41d0-a603-dc4598720ab4)
#### Old

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/2cba2c59-8661-4147-98f4-d2cb8a92f668)


### Seed
#### New

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/bb4780d6-0007-4c85-bc41-4a88f48c5dd8)
#### Old

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/12136af0-dd82-4c72-a86a-fc41d837c8d1)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

It was SO hard to use the tray one omg
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Fixed plant analyzer examine block styling
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->