Empty Gun Helper (#2954)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Makes all the current var edited and manually typed spawnwithmag = false
guns to be done with a helper you can slap under any gun you need it for

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The defines are alot quicker to do if you need it for something not
mapped, for mapping, it might still be better to var-edit.
At some point im probably going to detach the spawned mag for valid mag
types which this will make easier to change. Why?

![image](https://github.com/shiptest-ss13/Shiptest/assets/86381784/876ae8d8-931c-46f1-b46c-8606013dea9d)
Not hating on retlaw. this was the only way to do it with how we have it
set up rn


<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
refactor: spawning guns without mags are handled with a define
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>
Co-authored-by: Sun-Soaked <45698967+Sun-Soaked@users.noreply.github.com>